+++ CHANGELOG.md
@@ -11,10 +11,12 @@
 - Rename `configtest.LoadConfigFile` to `configtest.LoadConfigAndValidate` (#3306)
 - Replace `ExtensionCreateParams` with `ExtensionCreateSettings` (#3294)
 - Replace `ProcessorCreateParams` with `ProcessorCreateSettings`. (#3181)
+- Replace `ExporterCreateParams` with `ExporterCreateSettings` (#3164)
 
 ## 💡 Enhancements 💡
 
 - Add `doc.go` files to the consumer package and its subpackages (#3270)
+- Automate triggering of doc-update on release (#3234)
 - Enable Dependabot for Github Actions (#3312)
 
 ## v0.27.0 Beta